feat: udf onnx config placeholder
