+++ dozer-sql/src/pipeline/expression/tests/logical.rs
@@ -55,11 +55,13 @@ fn _test_bool_bool_and(bool1: bool, bool2: bool) {
     let row = ProcessorRecord::new();
     let l = Box::new(Literal(Field::Boolean(bool1)));
     let r = Box::new(Literal(Field::Boolean(bool2)));
-    assert!(matches!(
-        evaluate_and(&Schema::default(), &l, &r, &row)
-            .unwrap_or_else(|e| panic!("{}", e.to_string())),
-        Field::Boolean(_ans)
-    ));
+    assert!(
+        matches!(
+            evaluate_and(&Schema::default(), &l, &r, &row)
+                .unwrap_or_else(|e| panic!("{}", e.to_string())),
+            Field::Boolean(_ans)
+        )
+    );
 }
 
 fn _test_bool_null_and(f1: Field, f2: Field) {
